pyproject.toml
[tool.poetry]
name = "py-aiger-js"
version = "0.1.0"
repository = "https://github.com/mvcisback/py-aiger-js"
description = "A Python library for compiling AIGs to Javascript."
authors = ["Marcell Vazquez-Chanlatte <mvc@linux.com>"]
license = "MIT"
packages = [
    { include = "aiger_js" },
]

[tool.poetry.dependencies]
python = "^3.7"
pytest-flake8 = "^1.0.6"
pytest-xdist = "^1.32.0"
attrs = "^19.3.0"
funcy = "^1.14"
py-aiger = "^6.0.0"
py-aiger-ptltl = "^3.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
